% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_graficar.R
\name{distBarras}
\alias{distBarras}
\title{Para analizar y hacer una gráfica de barras de una pregunta en particular}
\usage{
distBarras(bd, pregunta, multi = F, titulo = "")
}
\arguments{
\item{bd}{base de datos}

\item{pregunta}{nombre de la columna a analizar}

\item{multi}{}

\item{titulo}{Título que se quiere en la gráfica}
}
\value{

}
\description{
Para analizar y hacer una gráfica de barras de una pregunta en particular
}
