% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_graficar.R
\name{graficar_burbujas}
\alias{graficar_burbujas}
\title{Para graficar burbujas}
\usage{
graficar_burbujas(bd, pregunta1, pregunta2)
}
\arguments{
\item{bd}{base de datos en la que vienen las respuestas del cuestionario}

\item{pregunta1}{nombre de la primera columna a graficar}

\item{pregunta2}{nombre de la columna que se graficará con la segunda}
}
\value{

}
\description{
Para graficar burbujas
}
