% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_graficar.R
\name{graficar_item}
\alias{graficar_item}
\title{Para graficar un respuesta en porcentaje con "bolitas" de colores por categoría}
\usage{
graficar_item(bd, pregunta, titulo = "")
}
\arguments{
\item{bd}{base de datos en la que vienen las respuestas del cuestionario}

\item{pregunta}{nombre de la columna que se va a graficar}

\item{titulo}{Título que aparecerá en la gráfica}
}
\value{

}
\description{
Para graficar un respuesta en porcentaje con "bolitas" de colores por categoría
}
