% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_graficar.R
\name{graficar_media_gauge}
\alias{graficar_media_gauge}
\title{Para graficar un gauge}
\usage{
graficar_media_gauge(
  bd,
  pregunta,
  grupo,
  colorPolos = c("#F78154", "#78B874"),
  titulo = "",
  tam = 10
)
}
\arguments{
\item{bd}{base de datos en la que vienen las respuestas del cuestionario}

\item{pregunta}{nombre de la columna que se va a graficar}

\item{grupo}{si si quiere agrupar por otra variable}

\item{colorPolos}{color de las líneas del gauge}

\item{titulo}{título que se mostrará en la gráfica}

\item{tam}{el tamaño de la la línea gauge máxima}
}
\value{

}
\description{
Para graficar un gauge
}
