% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_graficar.R
\name{graficar_paletas}
\alias{graficar_paletas}
\title{Función para hacer el procesamiento y la gráfica de lollipop o paleta}
\usage{
graficar_paletas(bd, pregunta, multi = F, titulo = "")
}
\arguments{
\item{bd}{base de datos en la que vienen las respuestas del cuestionario}

\item{pregunta}{nombre de la columna que se va a graficar}

\item{multi}{si se graficará una pregunta que es miltirrespuesta}

\item{titulo}{el título que se mostrará en la gráfica}
}
\value{

}
\description{
Función para hacer el procesamiento y la gráfica de lollipop o paleta
}
