% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_graficar.R
\name{graficar_piramide}
\alias{graficar_piramide}
\title{Para graficar pirámide poblacional o una pregunta con solo 2 tipos de repuestas agrupada}
\usage{
graficar_piramide(bd, pregunta, grupo, titulo = "")
}
\arguments{
\item{bd}{base de datos en la que vienen las respuestas del cuestionario}

\item{pregunta}{nombre de la columna que se va a graficar (usualmente edad)}

\item{grupo}{Nombre de la columna por la que se agrupará (usualmente sexo)}

\item{titulo}{Título que aparecerá en la gráfica}
}
\value{

}
\description{
Para graficar pirámide poblacional o una pregunta con solo 2 tipos de repuestas agrupada
}
