% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_graficar.R
\name{graficar_segmento}
\alias{graficar_segmento}
\title{Title}
\usage{
graficar_segmento(
  bd,
  pregunta,
  grupos,
  dominio = c(0, 10),
  polos = c("Polo 1", "Polo 2"),
  colorPolos = c("#F78154", "#78B874"),
  titulo = ""
)
}
\arguments{
\item{bd}{base de datos en la que vienen las respuestas del cuestionario}

\item{pregunta}{Variable numérica a graficar}

\item{grupos}{grupo por el que se calificó}

\item{dominio}{rango entre el mínimo y máximo que permitia la pregunta}

\item{polos}{Nombre que va en cada uno de los extremos de la gráfica}

\item{colorPolos}{color  de los polos}

\item{titulo}{Título que aparecerá en la gráfica}
}
\value{

}
\description{
Title
}
