% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaria_homologar_base.R
\name{leer_base}
\alias{leer_base}
\title{Para leer la base de datos de Survey Monkey}
\usage{
leer_base(archivo, dir_codigos = "data/", crear_codigo = T)
}
\arguments{
\item{archivo}{dirección de la base de datos del cuestionario}

\item{dir_codigos}{dirección en la que se escribirá la base de códigos}

\item{crear_codigo}{Si se quiere o no crear la base de códigos}
}
\value{
bd y codigos.csv
}
\description{
Para leer la base de datos de Survey Monkey
}
