% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaria_homologar_base.R
\name{nombrar_base}
\alias{nombrar_base}
\title{Para poner los nombres de códigos a la base de respuestas}
\usage{
nombrar_base(bd, bd_codigos)
}
\arguments{
\item{bd}{base de respuestas}

\item{bd_codigos}{base de datos de códigos con la columna "codigo"}
}
\value{
bd renombrada con codigos
}
\description{
Para poner los nombres de códigos a la base de respuestas
}
