% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_graficar.R
\name{transformar_multirespuesta}
\alias{transformar_multirespuesta}
\title{Title}
\usage{
transformar_multirespuesta(bd, grupo)
}
\arguments{
\item{bd}{base de datos del cuestionario}

\item{grupo}{patrón (string) de la respuesta multi, que debe ser el nombre de la variable}
}
\value{

}
\description{
Title
}
